pep8

+++ 100-clean_web_static.py
@@ -12,12 +12,13 @@ def do_clean(number=0):
     """deletes out-of-date archives"""
     target = sudo('ls -t ./AirBnB_Clone_V2/versions/').split()
     num = 0
-    if number == 0 or number == 1:
-            num = 1
-    elif number == 2:
-            num = 2
-    cl = target[num:]
-    for i in range(len(cl)):
-            local('rm f ./versions/{}'.format(cl[i]))
-            print(sudo("pwd"))
-            sudo('rm -rf /data/web_static/releases/{}'.format(cl[i].strip(".tgz")))
+    if len(target) > 1:
+        if number == 0 or number == 1:
+                num = 1
+        elif number == 2:
+                num = 2
+        cl = target[num:]
+        for i in range(len(cl)):
+                local('rm f ./versions/{}'.format(cl[i]))
+                print(sudo("pwd"))
+                sudo('rm -rf /data/web_static/releases/{}'.format(cl[i].strip(".tgz")))s
